requirements.txt
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
tensorflow-macos==2.15.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow==2.15.0; sys_platform != 'darwin' or platform_machine != 'arm64'
nltk==3.8.1
spacy==3.7.4
matplotlib==3.8.3
seaborn==0.13.2
Flask==3.0.2
joblib==1.3.2
imbalanced-learn==0.12.0
tqdm==4.66.2
moviepy
openai-whisper
Werkzeug
